refactor: DB query Dao로 관리
